refactor: Separate followed/following user queries from core
